COMMIT
Note structure for individual books/journals/resources with references and links to internal/external notes/books/journals/resources. With included reference to diagrams, tables, images, and tools.

TODO: More subsections and details, as well as examples of the work that can be or needs to be done in each one of the note-taking sections, more internal references and external references to note-taking/memorization methods as well as peer review methods and ways to learn at a more advanced level in a group. *Add more AI references and AI tools inside the notes as well as suggestions to what can be done and a TODO list.*